% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_trait_pars}
\alias{create_trait_pars}
\title{Create named list of trait state parameters}
\usage{
create_trait_pars(
  trans_rate,
  immig_rate2,
  ext_rate2,
  ana_rate2,
  clado_rate2,
  trans_rate2,
  M2
)
}
\value{
list of numerical values containing trait state parameters
}
\description{
Create named list of trait state parameters
}
\examples{
testit::assert(DAISIE:::are_trait_pars(
  create_trait_pars(
    trans_rate = 0.5,
    immig_rate2 = 0.1,
    ext_rate2 = 0.2,
    ana_rate2 = 0.3,
    clado_rate2 = 0.4,
    trans_rate2 = 0.5,
    M2 = 1000)) == TRUE)
}
